+++ Grains/AccountGrain.cs
@@ -1,16 +1,16 @@
 public class AccountGrain : Grain, IAccountGrain
 {
-    private decimal _balance;
+    private int _balance;
 
-    public Task<decimal> GetBalance() => Task.FromResult(_balance);
+    public Task<int> GetBalance() => Task.FromResult(_balance);
 
-    public Task Deposit(decimal amount)
+    public Task Deposit(int amount)
     {
         _balance += amount;
         return Task.CompletedTask;
     }
 
-    public Task<bool> Withdraw(decimal amount)
+    public Task<bool> Withdraw(int amount)
     {
         if (_balance >= amount)
         {
